+++ test/htmlwriter.jl
@@ -30,7 +30,9 @@ end
     @test isdir(HTMLWriter.ASSETS_THEMES)
 
     for theme in HTMLWriter.THEMES
-        @test isfile(joinpath(HTMLWriter.ASSETS_SASS, "$(theme).scss"))
+        # catppuccin-* themes are templated based on a common catppuccin.scss source file
+        scss = replace(theme, r"(?<=^catppuccin)-[a-z]+$" => "")
+        @test isfile(joinpath(HTMLWriter.ASSETS_SASS, "$(scss).scss"))
         @test isfile(joinpath(HTMLWriter.ASSETS_THEMES, "$(theme).css"))
     end